requirements.txt
kafka-python
fastapi[all]
python-dotenv
Faker
httpie